Fix: home index 변경 / video template 분리
