+++ src/components/templates/Video/index.ts
@@ -1,4 +1,4 @@
 export { default as VideoAbout } from "./VideoAbout";
 export { default as VideoForm } from "./VideoForm";
-export { default as VideoList } from "./VideoList";
+export { default as VideoList } from "./HomeVideoList";
 export { default as VideoPlayer } from "./VideoPlayer";
